[MLIR] Split InlinerConfig into seperate header and add pass overload with it
